+++ src/main/java/com/victorvalentim/zividomelive/_internal/render/core/CubemapRenderer.java
@@ -340,7 +340,7 @@ class CubemapRenderer implements PConstants {
         captureGraphics.beginDraw();
         try {
             captureGraphics.resetMatrix();
-            captureGraphics.background(0, 0);
+            captureGraphics.clear();
             configureCameraForFace(
                     captureGraphics,
                     orientation,